# ui/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "DeletedCounts": {
        "description": "Rows deleted by a project removal or a reset.",
        "properties": {
          "artifacts": {
            "format": "int64",
            "type": "integer"
          },
          "backfills": {
            "format": "int64",
            "type": "integer"
          },
          "events": {
            "format": "int64",
            "type": "integer"
          },
          "flows": {
            "format": "int64",
            "type": "integer"
          },
          "logs": {
            "format": "int64",
            "type": "integer"
          },
          "rules": {
            "format": "int64",
            "type": "integer"
          },
          "runs": {
            "format": "int64",
            "type": "integer"
          },
          "schedules": {
            "format": "int64",
            "type": "integer"
          },
          "task_runs": {
            "format": "int64",
            "type": "integer"
          },
          "variables": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "flows",
          "runs",
          "task_runs",
          "logs",
          "events",
          "artifacts",
          "schedules",
          "backfills",
          "rules",
          "variables"
        ],
        "type": "object"
      },
...
      "ProjectPreview": {
        "properties": {
          "active_runs": {
            "description": "Runs that are Pending, Running, Paused or Cancelling.",
            "format": "int64",
            "type": "integer"
          },
          "backfills": {
            "format": "int64",
            "type": "integer"
          },
          "events": {
            "format": "int64",
            "type": "integer"
          },
          "flows": {
            "format": "int64",
            "type": "integer"
          },
          "matching_rules": {
            "description": "Rules whose match names the project; they are kept.",
            "minimum": 0,
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "runs": {
            "format": "int64",
            "type": "integer"
          },
          "schedules": {
            "format": "int64",
            "type": "integer"
          },
          "served": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "flows",
          "runs",
          "schedules",
          "backfills",
          "events",
          "matching_rules",
          "served",
          "active_runs"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first pipeline orchestrator API",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "cereyan",
    "version": "1.14.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/{name}": {
      "delete": {
        "operationId": "delete_project",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedCounts"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": "The project is served, or a run of it is in progress"
          },
          "503": {
            "description": "The database is being reset"
          }
        },
        "tags": [
          "projects"
        ]
      },
      "get": {
        "operationId": "get_project",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPreview"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}